requirements.txt
tqdm
Pillow
matplotlib
numpy
pprint
pandas
requests-toolbelt
pma_python